% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.appDrugs}
\alias{get.appDrugs}
\alias{get.appDrugs-package}
\title{Get approved drugs for target.}
\usage{
get.appDrugs(x)
}
\arguments{
\item{x}{: ChEMBL target ID.}
}
\description{
This function retrieves a dataframe of all approved drug compounds from
ChEMBL database given a string of ChEMBL target ID.
}
\details{
get.appDrugs
}

