% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.bioactivity}
\alias{get.bioactivity}
\alias{get.bioactivity-package}
\title{Get Bioactivity Information for Compounds, Targets or Assays.}
\usage{
get.bioactivity(x, type = "compound")
}
\arguments{
\item{x}{: chemblid}

\item{type:}{\code{'compound'},\code{'target'},\code{'assay'}. Default is
\code{'compound'}.}
}
\description{
This method retrieves bioactivity information for a compound
across multiple targets/assays or else for a target across multiple compounds.
The function can also be used to retrieve all activities within a given assay.
In all cases, ChEMBL identifiers must be used.
}
\details{
get.bioactivity.
}
\examples{
get.bioactivity("CHEMBL12",type='compound')
get.bioactivity("CHEMBL240",type="target")
get.bioactivity("CHEMBL1217643",type='assay')
}

