% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.cmp.inf}
\alias{get.cmp.inf}
\alias{get.cmp.inf-package}
\title{Get compound information from ChEMBL}
\usage{
get.cmp.inf(x, type = "chemblid")
}
\arguments{
\item{x}{: String representing chemblid or standard InCHI key for the molecule.}

\item{type}{: For \code{get.compound}, one of \code{chemblid} or
\code{stdinchi} to indicate the nature of the molecule id.}
}
\description{
These functions allow one to retrieve compounds information from ChEMBL
compounds are identified either by a ChEMBL ID or by a standard InChI key.
}
\details{
get.cmp.inf
}
\examples{
\donttest{
#get information for chembl compound id
get.compound("CHEMBL12")

#get information for standard inchi
get.compound("QFFGVLORLPOAEC-SNVBAGLBSA-N",type='stdinchi')
}
}

