% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.cmp.sim}
\alias{get.cmp.sim}
\alias{get.cmp.sim-package}
\title{Retrive similar compounds from ChEMBL database.}
\usage{
get.cmp.sim(mol, sim = 70)
}
\arguments{
\item{mol}{: String representing smiles of the moelcule}

\item{sim}{: Integer representing for percentage of similarity
for the query compound and the database molecules. Values ranges
from 70 to 100.}
}
\description{
This function retrieves a dataframe of similar compounds
from ChEMBL database given a smiles string as query and also given a similarity score above 70.
}
\details{
get.cmp.sim
}
\examples{
# Get compound similarity with 70 percent or more
get.cmp.sim("c1ccc(cc1)CN(c2cc(ccc2[N+](=O)[O-])c3c(nc(nc3CC)N)N)C",sim=70)
# Get compound similarity with 80 percent or more
get.cmp.sim("c1ccc(cc1)CN(c2cc(ccc2[N+](=O)[O-])c3c(nc(nc3CC)N)N)C",sim=80)
}