% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.compound}
\alias{get.compound}
\alias{get.compound-package}
\title{Get compound information from ChEMBL}
\usage{
get.compound(x, type = "chemblid")
}
\arguments{
\item{x}{: String representing chemblid or standard InCHI key for the molecule.}

\item{type:}{For \code{get.compound}, one of \code{chemblid} or
\code{stdinchi} to #'indicate the nature of the molecule id.
For the case of \code{get.compound.list} valid types are \code{cansmi},
\code{substructure} and \code{similarity}.}
}
\description{
These functions allow one to retrieve compounds information from ChEMBL
compounds are identified either by a ChEMBL ID or by a standard InChI key.
}
\details{
get.compound
}

