% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.compound.substruct}
\alias{get.cmp.substruct}
\alias{get.compound.substruct}
\alias{get.compound.substruct-package}
\title{Get compound information from substructure query smiles.}
\usage{
get.cmp.substruct(mol)
}
\arguments{
\item{mol}{: String representing smiles of the moelcule}
}
\description{
This function retrieves a dataframe of all compounds from ChEMBL database
containing the substructure represented by the given Canonical SMILES and their
chemical properties.
}
\details{
get.cmp.substruct
}
\examples{
\donttest{
#get compounds by substructure
get.cmp.subsruct("CN(CCCN)c1cccc2ccccc12")
}
}

