% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.moa}
\alias{get.moa}
\alias{get.moa-package}
\title{Get mechanism of action}
\usage{
get.moa(x)
}
\arguments{
\item{x}{: Input string chemblid}
}
\description{
This function retrieves a data frame of compounds and its
mode of action for a compound (where compound is a drug) and drug targets.
}
\details{
get.moa
}
\examples{
\donttest{
# get moa of drug
get.moa("CHEMBL1642")
}
}

