% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.sAll.InCHIKey}
\alias{get.sAll.InCHIKey}
\alias{get.sAll.InCHIKey-package}
\title{Get all src_compound_ids.}
\usage{
get.sAll.InCHIKey(x)
}
\arguments{
\item{X}{: InCHI Key}
}
\description{
Get a list of all src_compound_ids (from all sources) which
have current AND obsolete assignments to a query InChIKey
}
\details{
get.sAll.InCHIKey
}

