% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.sAll.sid}
\alias{get.sAll.sid}
\alias{get.sAll.sid-package}
\title{Get the all source compound ids from another source compound id}
\usage{
get.sAll.sid(x, y)
}
\arguments{
\item{x}{: chemblid}

\item{y}{: source id}
}
\description{
Obtain a list of all src_compound_ids from all sources
(including BOTH current AND obsolete assignments) to the same structure
as a currently assigned query
src_compound_id.
}
\details{
get.sAll.sid
}
\examples{
get.sAll.sid("CHEMBL12","1")
get.sAll.sid("DB00789","2")
}

