% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.scid.sid}
\alias{get.scid.sid}
\alias{get.scid.sid-package}
\title{Get the source compound ids from another source compound id}
\usage{
get.scid.sid(x, y)
}
\arguments{
\item{x}{: Source compound id}

\item{y}{: Source id}
}
\description{
a list of all src_compound_ids from all sources which are
CURRENTLY assigned to the same structure as a currently assigned
query src_compound_id.
The output will include query src_compound_id if it is a valid src_compound_id
with a current assignment.
}
\details{
get.scid.sid
}
\examples{
get.scid.sid("CHEMBL12","1")
get.scid.sid("DB00789","2")
}

