% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.Src_cidAll.Src_id}
\alias{get.Src_cidAll.Src_id}
\alias{get.Src_cidAll.Src_id-package}
\alias{get.sid.InCHIKey}
\title{Get source compound ids}
\usage{
get.sid.InCHIKey(x)
}
\arguments{
\item{X}{: InCHI Key}
}
\description{
Obtain a list of src_compound_ids (from all sources) which
are CURRENTLY  assigned to a query InChI Key.
}
\details{
get.src_id.InCHIKey
}
\examples{
get.sid.InCHIKey("AAOVKJBEBIDNHE-UHFFFAOYSA-N")
get.sid.InCHIKey("BSYNRYMUTXBXSQ-UHFFFAOYSA-N")
}

