% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.src_id.InCHIKey}
\alias{get.sid.InCHIKey}
\alias{get.src_id.InCHIKey}
\alias{get.src_id.InCHIKey-package}
\title{Get source compound ids}
\usage{
get.sid.InCHIKey(x)
}
\arguments{
\item{x}{: Input string InCHI Key}
}
\description{
Obtain a list of src_compound_ids (from all sources) which
are CURRENTLY  assigned to a query InChI Key. Returns a list of data from
Unichem and ChEMBL databases.
}
\details{
get.src_id.InCHIKey
}
\examples{
\donttest{
# Get source compound ids from InCHIKey
get.sid.InCHIKey("AAOVKJBEBIDNHE-UHFFFAOYSA-N")

data<-get.sid.InCHIKey("BSYNRYMUTXBXSQ-UHFFFAOYSA-N")
# to get chembl data
data$Chem
to get Unichem data
data$Uni
}
}

