% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.SrcAll.obs}
\alias{get.SrcAll.obs}
\alias{get.SrcAll.obs-package}
\alias{get.sAll.obs}
\title{Get source compound id from obsolete source compound id}
\usage{
get.sAll.obs(x, y)
}
\arguments{
\item{x}{: Input string source compound id}

\item{y}{: Input integer to source id}
}
\description{
Get a list of all src_compound_ids from all sources with BOTH
current AND obsolete to the same structure with an obsolete assignment to the #' query src_compound_id.
}
\details{
get.sAll.obs
}
\examples{
\donttest{
#get for drugbank compound and source
get.sAll.obs("DB07699",2)
#get for chembl compound and source
get.sAll.obs("CHEMBL12",1)
}
}

