% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.structure}
\alias{get.structure}
\alias{get.structure-package}
\title{Get structure}
\usage{
get.structure(x, s = 1)
}
\arguments{
\item{x}{: Input string chemblid}

\item{s}{: Input integer source id (default is 1)}
}
\description{
Get structure(s) currently assigned to a query src_compound_id.
}
\details{
get.structure
}
\examples{
\donttest{
# Get Standard inhci and InCHIKey from drugbank compound and source
get.structure("DB00321",2)
# Using ChEMBL compound and source id
get.structure("CHEMBL1231",1)
}
}

