% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.targets}
\alias{get.targets}
\alias{get.targets-package}
\title{Get target information.}
\usage{
get.targets(x, type = "chemblid", org = NULL)
}
\arguments{
\item{x}{: chemblid}

\item{type:}{'chemblid' or 'uniprot'}

\item{org:}{Species name like "Homo sapiens","Plasmodium falciparum" and etc.}
}
\description{
This function retrieves the target information by chembl id and
uniprot id and also retrieves all the names of targets by organisms. When
\code{org="Homo sapiens"} subsets the data frame by organism homo sapiens and retrieves
all the Homo sapiens taregts
}
\details{
get.targets
}
\examples{
get.targets("CHEMBL1862",type='chemblid')

get.targets("Q13936",type='uniprot')

get.targets(org="Homo Sapiens")
}

