% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.url.sid}
\alias{get.url.sid}
\alias{get.url.sid-package}
\title{Get url for the query compound}
\usage{
get.url.sid(x, y, z)
}
\arguments{
\item{x}{: source compound id}

\item{y}{: source id}

\item{z}{: to source id}
}
\description{
Get a list of URLs for all src_compound_ids, from a specified
source .
}
\details{
get.url.sid
}

