% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.verbose.InCHIkey}
\alias{get.verbose.InCHIkey}
\alias{get.verbose.InCHIkey-package}
\title{Get all src_compound_ids to a query InChIKey}
\usage{
get.verbose.InCHIkey(x)
}
\arguments{
\item{x}{: Input string InCHI Key}
}
\description{
Returns a dataframe containing src_id (the src_id for this source),
src_url (the main home page of the source), name (the unique name for the source in
UniChem, always lower case), name_long (the full name of the source, as defined by the
source),name_label (A name for the source
suitable for use as a 'label' for the source within a web-page. Correct case setting
for source, and always less than 30 characters), description (a description of the
content of the source), base_id_url_available (an flag indicating whether this source
provides a valid base_id_url for creating cpd-specific links [1=yes, 0=no]),base_id_url
(the base url for constructing hyperlinks to this source [append an identifier from
this source to the end of this url to create a valid url to a specific page for this
cpd], unless aux_for_url=1), aux_for_url (A flag to indicate whether the aux_src field
should be used to create hyperlinks instead of the src_compound_id [1=yes, 0=no] ,
src_compound_id (a list of src_compound_ids from this source which are currently
assigned to the query InChIKey, aux_src (a list of src-compound_id keys mapping to
corresponding auxiliary data (url_id:value), for creating links if aux_for_url=1. Only
shown if aux_for_url=1).
}
\details{
get.verbose.InCHIkey
}
\examples{
\donttest{
# get for InCHIkey
get.verbose.InCHIkey("GZUITABIAKMVPG-UHFFFAOYSA-N")

get.verbose.InCHIkey("AAOVKJBEBIDNHE-UHFFFAOYSA-N")
}
}

