% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.mapping.full}
\alias{get.mapping.full}
\alias{get.mapping.full-package}
\title{Get full mapping between two sources.}
\usage{
get.mapping.full(x, y)
}
\arguments{
\item{x}{: Input integer source id}

\item{y}{: Input integer source id}
}
\description{
Obtain a full mapping between two sources. Uses only currently
assigned src_compound_ids from both sources.
}
\details{
get.mapping.full
}
\examples{
\donttest{
# Get full mapping of PDBe and ChEMBL
get.mapping.full(3,1)
# Get full mapping of ZINC and ChEMBL
get.mapping.full(9,1)
}
}

