% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.struc.all}
\alias{get.struc.all}
\alias{get.struc.all-package}
\title{Get Structures for source compound id}
\usage{
get.struc.all(x, s = 1)
}
\arguments{
\item{x}{: Input string chemblid}

\item{s}{: Input integer source id (default is 1)}
}
\description{
Get the standard InCHI and standard InCHI Key for the source
compound id
}
\details{
get.struc.all
}
\examples{
\donttest{
# Get all the structure information using ChEMBL id and source.
get.struc.all("CHEMBL1231",1)
#using drugbank id and source
get.structure("DB00321",2)
}
}

