% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{get.structure}
\alias{get.structure}
\alias{get.structure-package}
\title{Get structure}
\usage{
get.structure(x, s = 1)
}
\arguments{
\item{X}{: chemblid}

\item{s}{: source id (default is 1)}
}
\description{
Get structure(s) currently assigned to a query src_compound_id.
}
\details{
get.structure
}
\examples{
get.structure("DB00321","2")
get.structure("CHEMBL1231","1")
}

